+++ juriscraper/AbstractSite.py
@@ -1,8 +1,10 @@
 import hashlib
 import json
+import sys
 from datetime import date, datetime, timedelta
 
 import certifi
+import pytz
 import requests
 
 from juriscraper.lib.date_utils import (
@@ -78,6 +80,14 @@ class AbstractSite:
 
         # indicates whether the scraper should have results or not to raise an error
         self.should_have_results = False
+
+        # Default working hours, can be overridden by subclasses
+        self.working_hours = (0, 24)
+        # use print(pytz.all_timezones) to get a list of time zones
+        self.time_zone = "America/Los_Angeles"
+
+        # indicates the rate limit for the scraper, in seconds
+        self.rate_limit = 0
 
         # Sub-classed metadata
         self.court_id = None
@@ -354,8 +364,21 @@ class AbstractSite:
         """
         return get_html_parsed_text(text)
 
+    def is_within_working_hours(self) -> bool:
+        """
+        Checks if the current time is within the allowed working hours.
+
+        :return: True if within working hours, False otherwise.
+        """
+        now = datetime.now(pytz.timezone(self.time_zone)).time()
+        start, end = self.working_hours
+        return start <= now.hour < end
+
     def _download(self, request_dict=None):
         """Download the latest version of Site"""
+        if not self.is_within_working_hours():
+            raise sys.exit("Attempted to download outside of working hours.")
+
         if request_dict is None:
             request_dict = {}
         self.downloader_executed = True